% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{alignShapes}
\alias{alignShapes}
\title{Calculate distance offsets for a set of overlapping shapes}
\usage{
alignShapes(gtfs, from_stop, to_stop, crs, shape_ids = NULL)
}
\arguments{
\item{gtfs}{path to zipped GTFS feed}

\item{from_stop}{GTFS stop_id of stop that defines start of route segment.}

\item{to_stop}{GTFS stop_id of stop that defines end of route segment}

\item{crs}{local projection, default to UTM. Use a projection in meters.}

\item{shape_ids}{character vector of shape_ids,
one for each shape of interest in shapes.txt of GTFS. If NULL, will use all
shapes within 30 m of \code{from_stop} and \code{to_stop}.}
}
\value{
data.table of shape_id, from_dist, to_dist and stop_order 
To use, add from_dist to avl_dist_traveled for each shape_id in a matched AVL
dataset.
}
\description{
Calculate distance offsets for a set of overlapping shapes
}
\examples{
gtfs <- system.file('extdata', 'gtfs.zip', package = 'speedRT')
shape_offsets <- alignShapes(gtfs, from_stop = 13337, to_stop = 13320)
}
