% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.TransitTime}
\alias{as.TransitTime}
\title{Convert character time to ITime, 28 hour clock}
\usage{
as.TransitTime(x)
}
\arguments{
\item{x}{time represented as a character string}
}
\value{
seconds from midnight with ITime class, service continuing past
midnight will continue to count seconds from previous day.
}
\description{
Convert character time to ITime, 28 hour clock
}
\examples{
as.TransitTime(c('1:00', '25:00'))
}
