% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{distanceAlongShape}
\alias{distanceAlongShape}
\title{How far along a shape linestring is a point?}
\usage{
distanceAlongShape(lon, lat, shape_id., srid, shape_geo)
}
\arguments{
\item{lon}{longitude WGS84 longlat}

\item{lat}{latitude WGS84 longlat}

\item{shape_id.}{GTFS shape identifier}

\item{srid}{local projection}

\item{shape_geo}{data.table with shape_id and sfc_LINESTRING geom columns}
}
\value{
list of distance along shape, latitude and longitude of matched points
}
\description{
How far along a shape linestring is a point?
}
