% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_positions.R
\name{fetch_feed}
\alias{fetch_feed}
\title{Download protobuf files}
\usage{
fetch_feed(feed, output)
}
\arguments{
\item{feed}{URL for protobuf file.}

\item{output}{path for download (directory).}
}
\value{
response content
}
\description{
Download protobuf files
}
