% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{filterMatches}
\alias{filterMatches}
\title{Filter matched locations so points travel forward along shape}
\usage{
filterMatches(avl_matches, max_speed = Inf)
}
\arguments{
\item{avl_matches}{vehicle positions matched to route shapes.
See \code{\link{matchAVL}} for matching.}

\item{max_speed}{maximum expected speed (in units of projection per second,
typically m/s). 
Observations over max_speed will be dropped and speeds recalculated.}
}
\value{
data.table of filtered values in avl_matches, with dist_to_next and
  mps (meters per second) columns.
}
\description{
Removes vehicle positions that result in negative distance traveled between
observations. 
When "odometer" values are available, also excludes positions with odometer
values that decrease for increasing timestamp values.
}
\examples{
\dontrun{
avl <- readVehiclePosition(system.file('extdata', 'vehiclePositions', package = 'speedRT'))
matched <- matchAVL(avl, gtfs = system.file('extdata', 'gtfs.zip', package = 'speedRT'))
filtered <- filterMatches(matched)
}
}
