% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{inferUTM}
\alias{inferUTM}
\title{Infer UTM Zone from lon-lat or sf POINT}
\usage{
inferUTM(x)
}
\arguments{
\item{x}{either a long-lat pair \code{c(-93, 45)} or an sf POINT object.}
}
\value{
epsg of best fit UTM zone.
}
\description{
Infer UTM Zone from lon-lat or sf POINT
}
