% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedLine.R
\name{interpolateLatLon}
\alias{interpolateLatLon}
\title{Interpolate locations along a shape}
\usage{
interpolateLatLon(d, id, shapes, crs)
}
\arguments{
\item{d}{distance between interpolated points, in units of \code{crs}}

\item{id}{valid GTFS shape_id}

\item{shapes}{GTFS shapes table, see \code{\link{readGTFS}}.}

\item{crs}{a projected coordinate reference system, see \code{\link{st_crs}}.}
}
\value{
a list of vectors of longitude and latitude for each interpolated 
point.
}
\description{
Requires packages sp and rgeos.
}
