% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_speedapp.R
\name{launch_speedapp}
\alias{launch_speedapp}
\title{Launch the Shiny Speed app}
\usage{
launch_speedapp(sao = NULL, rstudio = Sys.getenv("RSTUDIO") == "1",
  ...)
}
\arguments{
\item{sao}{speedapp object}

\item{rstudio}{launch in rstudio viewer instead of web browser?}

\item{...}{passed to shiny::runApp}
}
\description{
Launch the Shiny Speed app
}
