% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_positions.R
\name{logProtobufFeed}
\alias{logProtobufFeed}
\title{Log Protobuf Feeds including GTFS-RT Vehicle Positions}
\usage{
logProtobufFeed(feed, refresh = 10L, duration = 10L, output = NULL)
}
\arguments{
\item{feed}{URL of protobuf file.}

\item{refresh}{minimum time in seconds between queries of the protobuf feed.}

\item{duration}{total time in seconds to query feed.}

\item{output}{location to save files, if \code{NULL} will save to a tempfile.}
}
\value{
path to saved protobuf files
}
\description{
Log Protobuf Feeds including GTFS-RT Vehicle Positions
}
