% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_positions.R
\name{logVehiclePositions}
\alias{logVehiclePositions}
\title{Log GTFS-RT Vehicle Positions}
\usage{
logVehiclePositions(feed, refresh = 10L, duration = 10L,
  output = NULL)
}
\arguments{
\item{feed}{URL for GTFS-Realtime feed with VehiclePosition messages.}

\item{refresh}{minimum time in seconds between queries of the GTFS-Realtime feed.}

\item{duration}{total time in seconds to query feed.}

\item{output}{location to save files, if \code{NULL} will save to a tempfile.}
}
\value{
path to saved realtime files
}
\description{
Log GTFS-RT Vehicle Positions
}
