% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{lookupService}
\alias{lookupService}
\title{Infer service names from GTFS calendar}
\usage{
lookupService(gtfs)
}
\arguments{
\item{gtfs}{path to the zipped GTFS feed.}
}
\value{
data.table linking service_id to a descriptive service_name
}
\description{
Infer service names from GTFS calendar
}
\examples{
gtfs <- system.file('extdata', 'gtfs.zip', package = 'speedRT')
service_lu <- lookupService(gtfs)
}
