% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{matchAVL}
\alias{matchAVL}
\title{Match AVL positions to route shapes}
\usage{
matchAVL(avl, gtfs, crs = inferUTM(avl[1, c(longitude, latitude)]),
  tz = Sys.timezone(), within_dist = 30)
}
\arguments{
\item{avl}{a data.table. Required fields: trip_id, start_date, latitude, 
longitude, timestamp, vehicle_id. 
Other fields are currently ignored, though bearing and odometer could be 
used eventually.}

\item{gtfs}{path to a zipped GTFS feed with identifier that match \code{avl}.}

\item{crs}{local projection, default to UTM. Use a projection in meters.}

\item{tz}{timezone of observations}

\item{within_dist}{distance in units of CRS (typically meters) from shape. 
Vehicle positions outside of distance will be dropped from the result.}
}
\value{
data.table with trip_id, route_id, latitude, longitude, timestamp, 
  vehicle_id, start_date, match_lat, match_lon, and avl_dist_traveled.
  The data.table will also include any additional fields in \code{avl}.
}
\description{
Match AVL positions to route shapes
}
\examples{
\dontrun{
avl <- readVehiclePosition(system.file('extdata', 'vehiclePositions', package = 'speedRT'))	
matched <- matchAVL(avl, gtfs = system.file('extdata', 'gtfs.zip', package = 'speedRT'))
}
}
