% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpeedDist}
\alias{plotSpeedDist}
\title{Plot speed by distance traveled}
\usage{
plotSpeedDist(speed, compare = "None", alpha = 0)
}
\arguments{
\item{speed}{avl speeds interpolated along shapes,
see \code{\link{speedOnLine}}.}

\item{compare}{grouping for speed comparison, one of 'DOW', 'TOD',
'date_range', 'route_short_name', or 'None'.}

\item{alpha}{lower bound of confidence interval, 
e.g., 0.05 results in an interval of 0.05-0.95 or 90\%.}
}
\value{
a ggplot object.
}
\description{
Plot speed by distance traveled
}
