% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSpeedHistogram}
\alias{plotSpeedHistogram}
\title{Plot a speed density}
\usage{
plotSpeedHistogram(avl, compare = "None")
}
\arguments{
\item{avl}{avl speeds matched to shapes, see \code{\link{filterMatches}}.}

\item{compare}{grouping for speed comparison, one of 'DOW', 'TOD', 
'date_range', 'route_short_name', or 'None'.}
}
\value{
ggplot object
}
\description{
Plot a speed density
}
