% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchPoints.R
\name{pointsWithin}
\alias{pointsWithin}
\title{Are points within a distance of a shape?}
\usage{
pointsWithin(lon, lat, shape_id., srid, dist = 30, shape_geo)
}
\arguments{
\item{lon}{longitude WGS84 longlat}

\item{lat}{latitude WGS84 longlat}

\item{shape_id.}{GTFS shape identifier}

\item{srid}{local projection}

\item{dist}{distance in units of local projects}

\item{shape_geo}{data.table with shape_id and sfc_LINESTRING geom columns}
}
\value{
logical same length as lon, indicating if points defined by lon, lat 
   are within dist of shape defined by shape_id.
}
\description{
Are points within a distance of a shape?
}
