% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_positions.R
\name{readVehiclePosition}
\alias{readVehiclePosition}
\title{Read VehiclePosition messages from GTFS-realtime ProtoBuf file}
\usage{
readVehiclePosition(descriptor)
}
\arguments{
\item{descriptor}{location of protobuf file or raw vector.}
}
\value{
a data.table with fields: trip_id, start_date, route_id, lat, lon, 
  bearing, stop_sequence, status, timestamp, vehicle_id
}
\description{
Read VehiclePosition messages from GTFS-realtime ProtoBuf file
}
