% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{shapesToLinestring}
\alias{shapesToLinestring}
\title{Create linestrings from GTFS Shapes}
\usage{
shapesToLinestring(gtfs)
}
\arguments{
\item{gtfs}{location of zipped GTFS feed}
}
\value{
data.table with shape_id and geom columns
}
\description{
Create linestrings from GTFS Shapes
}
