% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedLine.R
\name{speedOnLine}
\alias{speedOnLine}
\title{Interpolate speed along a route shape}
\usage{
speedOnLine(avl, by_dist = 1, shapes, crs)
}
\arguments{
\item{avl}{speeds data.table}

\item{by_dist}{distance between gridded points, in units of crs.}

\item{shapes}{GTFS shapes table, see \code{\link{readGTFS}}.}

\item{crs}{a projected coordinate reference system, see \code{\link{st_crs}}.}
}
\value{
a data.table of speeds with imputed lonlat 
(\code{lon_imp} and \code{lat_imp}) for distances \code{by_dist} apart.
}
\description{
Requires packages sp and rgeos.
}
