% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedRT-package.R
\docType{package}
\name{speedRT-package}
\alias{speedRT}
\alias{speedRT-package}
\title{speedRT: Visualize speed from GTFS-RT data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/metrotransit/speedrt/issues}
}

}
\author{
\strong{Maintainer}: Joey Reid \email{joey.reid@metrotransit.org} (0000-0001-6905-5235)

Other contributors:
\itemize{
  \item  MetroTransit Strategic Initiatives [contributor]
  \item  Metropolitan Council [copyright holder]
}

}
\keyword{internal}
