% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{speedToTT}
\alias{speedToTT}
\title{Plot cumulative travel time by distance}
\usage{
speedToTT(speed, compare = "None")
}
\arguments{
\item{speed}{avl speeds interpolated along shapes,
see \code{\link{speedOnLine}}.}

\item{compare}{grouping for speed comparison, one of 'DOW', 'TOD', 
'date_range', 'route_short_name', or 'None'.}
}
\value{
named list of ggplot objects, one for each member of \code{compare}.
}
\description{
Plot cumulative travel time by distance
}
