% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{writeGTFS}
\alias{writeGTFS}
\title{Write tables to a GTFS feed}
\usage{
writeGTFS(x, name = c("agency", "stops", "routes", "trips", "stop_times",
  "calendar", "calendar_dates", "fare_attributes", "fare_rules", "shapes",
  "frequencies", "transfers", "feed_info"), gtfs)
}
\arguments{
\item{x}{data.table to write}

\item{name}{name of the table to read, without ".txt" extension.
Current tables include 'agency', 'stops', 'routes', 'trips',
'stop_times', 'calendar', 'calendar_dates', 'fare_attributes', 
'fare_rules', 'shapes', 'frequencies', 'transfers', 'feed_info'.
See the \href{https://gtfs.org/reference/static/}{GTFS Static Reference} for a full list.}

\item{gtfs}{location of zipped GTFS feed}
}
\value{
0 if successful
}
\description{
Write tables to a GTFS feed
}
\examples{
gtfs <- system.file('extdata', 'gtfs.zip', package = 'speedRT')
routes <- readGTFS('routes', gtfs)
routes[1, `:=` (route_short_name = 'Metro 2')]
writeGTFS(routes, 'routes', file.path(tempdir(), 'gtfs_test.zip'))
}
