% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_positions.R
\name{fetch_feed}
\alias{fetch_feed}
\title{Log GTFS-RT Vehicle Positions}
\usage{
fetch_feed(feed, output)
}
\arguments{
\item{feed}{URL for GTFS-Realtime feed with VehiclePosition messages.}

\item{output}{path for download}
}
\value{
response content
}
\description{
Log GTFS-RT Vehicle Positions
}
